chore: improve godoc
